Fix VS 2010 build

+++ yaf_opengl_engine/src/Main.cpp
@@ -206,7 +206,7 @@ YafScene* ParseYafFile(const std::string& file)
                 node->AddTransform(transform);
             }
             else
-                throw YafParsingException("Transform " + (*t)->ValueStr() + " is not recognized: line: " + std::to_string((*t)->Row()) + " col: " + std::to_string((*t)->Column()));
+                throw YafParsingException("Transform " + (*t)->ValueStr() + " is not recognized: line: " + std::to_string((long long)(*t)->Row()) + " col: " + std::to_string((long long)(*t)->Column()));
         }
 
         if (auto appRefElement = GetChildren(*n, "appearanceref", "graph node", false))
